+++ midap/tracking/deltav1_tracking.py
@@ -1,7 +1,8 @@
-from ..networks.deltav2 import unet_track
+from ..networks.deltav1 import unet_track
 from .base_tracking import Tracking
+import tensorflow as tf
 
-class DeltaV2Tracking(Tracking):
+class DeltaV1Tracking(Tracking):
     """
     A class for cell tracking using the U-Net Delta V2 model
 
@@ -47,4 +48,13 @@ class DeltaV2Tracking(Tracking):
             Array containing the constant input (whole raw image and segmentation image) per time frame.
         """
 
-        self.model = unet_track(self.model_weights, self.input_size)
+        # we get the model
+        model = unet_track(self.input_size, constant_input=None)
+        model.load_weights("../model_weights/model_weights_tracking/unet_moma_track_multisets.hdf5")
+
+        # now we create a Delta2 conform model, this is similiar to what was done before DeltaV2
+        inputs = tf.keras.layers.Input(shape=self.input_size, dtype='float32')
+        intermediate = model(inputs)
+        outputs = tf.reduce_sum(tf.where(intermediate[...,:2] > 0.8, 1.0, 0.0), keepdims=True, axis=-1)
+
+        self.model = tf.keras.Model(inputs=inputs, outputs=outputs)